% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubercv.R
\name{cv.hdhuber}
\alias{cv.hdhuber}
\title{Cross-validation for Selecting the Tuning Parameter in Penalized Huber Regression}
\usage{
cv.hdhuber(x, y, lambda = NULL, delta, nfolds = 5L, foldid, ...)
}
\arguments{
\item{x}{A numerical matrix with \eqn{n} rows (observations) and \eqn{p} columns (variables).}

\item{y}{Response variable.}

\item{lambda}{Optional; a user-supplied sequence of \code{lambda} values. If \code{NULL}, 
\code{\link{hdhuber}} selects its own sequence.}

\item{delta}{(\code{delta}) used in the loss function.}

\item{nfolds}{Number of folds for cross-validation. Defaults to 5.}

\item{foldid}{Optional vector specifying the indices of observations in each fold.
If provided, it overrides \code{nfolds}.}

\item{...}{Additional arguments passed to \code{\link{hdhuber}}.}
}
\value{
An object with S3 class \code{cv.hdhuber} consisting of
  \item{lambda}{Candidate \code{lambda} values.}
  \item{cvm}{Mean cross-validation error.}
  \item{cvsd}{Standard error of the mean cross-validation error.}
  \item{cvup}{Upper confidence curve: \code{cvm} + \code{cvsd}.}
  \item{cvlo}{Lower confidence curve: \code{cvm} - \code{cvsd}.}
  \item{lambda.min}{\code{lambda} achieving the minimum cross-validation error.}
  \item{lambda.1se}{Largest \code{lambda} within one standard error of the minimum error.}
  \item{cv.min}{Cross-validation error at \code{lambda.min}.}
  \item{cv.1se}{Cross-validation error at \code{lambda.1se}.}
  \item{hdhuber.fit}{a fitted \code{\link{hdhuber}} object for the full data.}
  \item{nzero}{Number of non-zero coefficients at each \code{lambda}.}
}
\description{
Performs k-fold cross-validation for \code{\link{hdhuber}}.
}
\details{
This function computes the average cross-validation error and provides the standard error.
}
\examples{
set.seed(315)
n <- 100
p <- 400
x <- matrix(data = rnorm(n * p, mean = 0, sd = 1), nrow = n, ncol = p)
beta_star <- c(c(2, 1.5, 0.8, 1, 1.75, 0.75, 0.3), rep(0, (p - 7)))
eps <- rnorm(n, mean = 0, sd = 1)
y <- x \%*\% beta_star + eps
delta = 0.5
cv.fit <- cv.hdhuber(x = x, y = y, delta = delta)
}
\keyword{models}
\keyword{regression}
